+++ cmd/paramgen/internal/testdata/basic/go.mod
