+++ cmd/paramgen/internal/testdata/basic/specs.go
