+++ cmd/paramgen/internal/testdata/complex/global.go
