+++ cmd/paramgen/internal/testdata/complex/internal/global.go
