+++ cmd/paramgen/internal/testdata/complex/specs.go
