+++ cmd/paramgen/internal/testdata/complex/specs_test.go
