+++ cmd/paramgen/internal/testdata/invalid1/specs.go
