+++ cmd/paramgen/internal/testdata/invalid2/go.mod
