+++ cmd/paramgen/internal/testdata/invalid2/specs.go
